update variable from int to string

+++ mobile-ui/src/main/java/com/thomaskioko/paybillmanager/mobile/ui/fragment/ConfirmBillFragment.kt
@@ -90,7 +90,7 @@ class ConfirmBillFragment : Fragment(), Injectable, Step {
                 payBillNumber,
                 accountNumber,
                 amount,
-                categoryId.toInt(),
+                categoryId,
                 OffsetDateTime.now().toEpochSecond()
         )
